chore: move SeatAdapter, SeatActivity, update manifest [Datto]
